use correct type hints for audio transcriptions
